// packages/ds-table/src/table/table.types.ts
type TableFunction<T> = (item: any) => T;
type TableFunctionRenderValue<T> = (
    item: any,
    column: TableColumn,
    rowIndex: number,
    colIndex: number
) => T;
// ...
export interface TableColumn {
    name: string; // Column name
    key: string; // Column key (unique identifier)
    type?: TableColumnType; // Column type (number, currency, date, etc.)
    url?: string | TableFunction<string>; // URL or URL-generating function for the column
    secondaryUrl?: string | TableFunction<string>; // Secondary URL or URL-generating function for the column
    visible?: boolean; // Whether the column is visible
    sortable?: boolean; // Whether the column is sortable
    ellipse?: number | boolean; // Number of characters to display or whether to use ellipsis
    className?: string | TableFunction<string>; // CSS class or class-generating function for the column
    renderValue?: TableFunctionRenderValue<ReactNode>; // Function for rendering the column value
    action?: (item: any) => void; // Ading action in column
    enableDbSort?: boolean; // Whether to enable sorting by the column in the database
    doubleConfirmation?: boolean;
    method?: string;
    dynamicStatus?: (
        status?: boolean,
        item?: ObjectDto
    ) => 'activate' | 'activate_badge';
    statusBadgeSize?: 'xs' | 'sm' | 'md' | 'lg' | 'normal' | 'polaris-normal';
    activateId?: string;
    disableRightClick?: boolean; // this will allow to disableRight click in the field
    disableAggregation?: boolean; // this will allow to disable aggregation in the field
    attributes?: ObjectDto;
    headerClassName?: string;
    getRightClickCustomQuery?: () => {
        value: string;
        field: string;
    };
    truncateClass?: string;
}

export interface SortingType {
    column: string;
    order: 'asc' | 'desc';
}
// ...
export type TableColumnType =
    | 'text'
    | 'reference_select'
    | 'number'
    | 'currency'
    | 'date'
    | 'date_time'
    | 'currency'
    | 'boolean'
    | 'activate'
    | 'activate_badge'
    | 'currency_acc'
    | 'dualistic'
    | 'date_lateral';